refactor(infra): renames context folder to contexts
