+++ template/src/contexts/AuthContext.tsx
