+++ template/src/index.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { ChakraProvider } from '@chakra-ui/react';
-import { AuthProvider } from './context/AuthContext';
+import { AuthProvider } from './contexts/AuthContext';
 import { makeServer } from './server';
 import theme from './styles/theme';
